Move boingboing to _examples/, mpsquic to pkg/
Comment out quictrace (not available in old quic-go, need some work to
update this here (because of h2quic)).
